change off_t to size_t on the advice of Glynn
git-svn-id: https://svn.osgeo.org/grass/grass/trunk@48258 15284696-431f-4ddb-bdfa-cd5b030d7da7

+++ raster/r.univar/globals.h
@@ -41,7 +41,7 @@ typedef struct
     CELL *cell_array;
     int map_type;
     void *nextp;
-    off_t n_alloc;
+    size_t n_alloc;
     int first;
 } univar_stat;
